Bug fixing
Previously the program would get the download by pressing a varity of buttons now however it just request the download via a link and some cookies, this lowered the chance of failure and improved efficacy. The older version had a problem where it would fail to download bring up a 500 error.